service: Add EmailService for async notifications
Added an EmailService for asynchronous notifications
to the users whenever their accounts get created.

The EmailService runs on a separate thread, gets
the mail from a Queue using the wait-notify mechanism
and emulates their sending. The mails are added to the
Queue by the main thread, by calling sendNotificationEmail().

At the end, the main thread closes the EmailService,
waits for the secondary thread to send all emails and
exists gracefully.

Signed-off-by: Andrei Cherechesu <andrei.cherechesu@gmail.com>